Expose an API to startServer.
